Update README.md
Added Sophos underneath Threatlocker because it's the only other set of instructions with available cert exemption. Waiting on some final direction from Sophos so wait on a comment from me before approving. This entire section needs to be reorganized but I'll wait to do that in discourse at a future time